requirements.txt
streamlit
requests
exa_py
openai
linkup-sdk
google-generativeai
python-dotenv
pydantic
psycopg2-binary